src/dataharvesting/requirements.txt
async-generator==1.10
attrs==22.2.0
black @ git+https://github.com/psf/black@ef6e079901d53a42dfae4ab10b081ce7a73a47b5
certifi==2022.12.7
click==8.1.3
exceptiongroup==1.1.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
idna==3.4
importlib-metadata==6.1.0
mypy-extensions==1.0.0
numpy==1.21.6
outcome==1.2.0
packaging==23.0
pandas==1.3.5
pathspec==0.11.1
platformdirs==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
selenium==4.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
trio==0.22.0
trio-websocket==0.10.2
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.26.15
wsproto==1.2.0
zipp==3.15.0
